requirements.txt
python-instagram==1.3.2 -e git+git://github.com/philippze/python-instagram.git#egg=python-instagram
python-dotenv==0.2.0
flask==0.10.1
python-twitter
